refactor: jdbc entity mapping 책임 분리
